refactor: add X11 Linux UI backend
