+++ src/ui.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#ifdef _WIN32
+#include "ui_windows.hpp"
+#else
+#include "ui_linux.hpp"
+#endif
