+++ src/ui_windows_settings_dialog_ids.hpp
@@ -1,0 +1,42 @@
+// Included by ui_windows.hpp inside namespace settings_dlg.
+
+constexpr int IDC_POMO_WORK    = 301;
+constexpr int IDC_POMO_SHORT   = 302;
+constexpr int IDC_POMO_LONG    = 303;
+constexpr int IDC_LBL_WORK     = 304;
+constexpr int IDC_LBL_SHORT    = 305;
+constexpr int IDC_LBL_LONG     = 306;
+constexpr int IDC_MIN_WORK     = 307;
+constexpr int IDC_MIN_SHORT    = 308;
+constexpr int IDC_MIN_LONG     = 309;
+constexpr int IDC_POMO_CADENCE = 310;
+constexpr int IDC_LBL_CADENCE  = 311;
+constexpr int IDC_MIN_CADENCE  = 312;
+constexpr int IDC_SET_OK       = 313;
+constexpr int IDC_SET_CANCEL   = 314;
+constexpr int IDC_PRESET_ED0   = 320;
+constexpr int IDC_PRESET_ED1   = 321;
+constexpr int IDC_PRESET_ED2   = 322;
+constexpr int IDC_PRESET_ED3   = 323;
+constexpr int IDC_PRESET_ED4   = 324;
+constexpr int IDC_PRESET_MIN0  = 325;
+constexpr int IDC_PRESET_MIN1  = 326;
+constexpr int IDC_PRESET_MIN2  = 327;
+constexpr int IDC_PRESET_MIN3  = 328;
+constexpr int IDC_PRESET_MIN4  = 329;
+constexpr int IDC_CLOCK_COMBO  = 330;
+
+constexpr WORD CLS_BUTTON   = 0x0080;
+constexpr WORD CLS_EDIT     = 0x0081;
+constexpr WORD CLS_STATIC   = 0x0082;
+constexpr WORD CLS_COMBOBOX = 0x0085;
+
+constexpr short DLG_W = 240, DLG_H = 156;
+constexpr short SIDEBAR_W = 62;
+constexpr short TITLE_H = 18;
+
+constexpr int TAB_APPEARANCE = 0;
+constexpr int TAB_CLOCK      = 1;
+constexpr int TAB_POMODORO   = 2;
+constexpr int TAB_TIMERS     = 3;
+constexpr int TAB_COUNT      = 4;
